refactor(web/applications): move search results to root folder (boas-740)
