+++ apps/web/src/server/applications/applications.router.js
@@ -6,7 +6,7 @@ import * as guards from './applications.guards.js';
 import * as locals from './applications.locals.js';
 import applicationsCaseRouter from './pages/case/applications-case.router.js';
 import applicationsCreateRouter from './pages/create-new-case/applications-create.router.js';
-import applicationsSearchRouter from './pages/search-results/applications-search.router.js';
+import applicationsSearchRouter from './search-results/applications-search.router.js';
 
 const router = createRouter();
 const domainRouter = createRouter({ mergeParams: true });
