+++ apps/web/src/server/applications/search-results/application-search.service.js
@@ -1,8 +1,8 @@
-import { post } from '../../../lib/request.js';
+import { post } from '../../lib/request.js';
 
 /** @typedef {import('./applications-search.types.js').ApplicationsSearchResultsBody} ApplicationsSearchResultsBody */
-/** @typedef {import('../../applications.types').Case} Case */
-/** @typedef {import('../../applications.types.js').PaginatedResponse<Case>} PaginatedApplicationsResponse */
+/** @typedef {import('../applications.types').Case} Case */
+/** @typedef {import('../applications.types.js').PaginatedResponse<Case>} PaginatedApplicationsResponse */
 
 /**
  * @param {ApplicationsSearchResultsBody} payload
