+++ apps/web/src/server/applications/search-results/application-search.validators.js
