+++ apps/web/src/server/applications/search-results/applications-search.types.d.ts
